refactor the calculateAge func to utils folder
